+++ lib/easy_talk/schema_definition.rb
@@ -3,12 +3,14 @@
 require_relative 'keywords'
 
 module EasyTalk
+  class InvalidPropertyNameError < StandardError; end
   #
   #= EasyTalk \SchemaDefinition
   # SchemaDefinition provides the methods for defining a schema within the define_schema block.
   # The @schema is a hash that contains the unvalidated schema definition for the model.
   # A SchemaDefinition instanace is the passed to the Builder.build_schema method to validate and compile the schema.
   class SchemaDefinition
+
     extend T::Sig
     extend T::AnyOf
     extend T::OneOf
@@ -35,15 +37,27 @@ module EasyTalk
     sig do
       params(name: T.any(Symbol, String), type: T.untyped, constraints: T.untyped, blk: T.nilable(T.proc.void)).void
     end
-    def property(name, type, **constraints, &blk)
+    def property(name, type, constraints = {}, &blk)
+      validate_property_name(name)
       @schema[:properties] ||= {}
 
       if block_given?
-        property_schema = SchemaDefinition.new(name, constraints)
+        property_schema = SchemaDefinition.new(name)
         property_schema.instance_eval(&blk)
-        @schema[:properties][name] = property_schema
+
+        @schema[:properties][name] = {
+          type:,
+          constraints:,
+          properties: property_schema
+        }
       else
         @schema[:properties][name] = { type:, constraints: }
+      end
+    end
+
+    def validate_property_name(name)
+      unless name.to_s.match?(/^[A-Za-z_][A-Za-z0-9_]*$/)
+        raise InvalidPropertyNameError, "Invalid property name '#{name}'. Must start with letter/underscore and contain only letters, numbers, underscores"
       end
     end