+++ lib/easy_talk/tools/function_builder.rb
@@ -5,23 +5,35 @@ module EasyTalk
     # FunctionBuilder is a module that builds a hash with the function type and function details.
     # The return value is typically passed as argument to LLM function calling APIs.
     module FunctionBuilder
-      # Creates a new function object based on the given model.
-      #
-      # @param [Model] model The EasyTalk model containing the function details.
-      # @return [Hash] The function object.
-      def self.new(model)
-        {
-          type: 'function',
-          function: {
-            name: model.function_name,
-            description: generate_description(model),
-            parameters: model.json_schema
+      class << self
+        # Creates a new function object based on the given model.
+        #
+        # @param [Model] model The EasyTalk model containing the function details.
+        # @return [Hash] The function object.
+        def new(model)
+          {
+            type: 'function',
+            function: {
+              name: generate_function_name(model),
+              description: generate_function_description(model),
+              parameters: model.json_schema
+            }
           }
-        }
-      end
+        end
 
-      def self.generate_description(model)
-        "Correctly extracted `#{model.name}` with all the required parameters with correct types"
+        def generate_function_name(model)
+          model.schema.fetch(:title, model.name)
+        end
+
+        def generate_function_description(model)
+          if model.respond_to?(:instructions)
+            raise Instructor::Error, 'The instructions must be a string' unless model.instructions.is_a?(String)
+
+            model.instructions
+          else
+            "Correctly extracted `#{model.name}` with all the required parameters and correct types."
+          end
+        end
       end
     end
   end